// report-ng/app-react/src/model/report-model/framework_pb.ts
export interface ErrorContext {
  /**
   * string readable_error_message = 1 [deprecated = true];
   *    string additional_error_message = 2 [deprecated = true];
   *    StackTrace stack_trace = 3  [deprecated = true];
   *    string error_fingerprint = 6 [deprecated = true];
   */
  scriptSource?:
    | ScriptSource
    | undefined;
  /**
   * ScriptSource execution_object_source = 8 [deprecated = true];
   *
   * @deprecated
   */
  ticketId?:
    | string
    | undefined;
  /** @deprecated */
  description?: string | undefined;
  stackTrace?: StackTraceCause[] | undefined;
  optional?: boolean | undefined;
  id?: string | undefined;
}
// ...
export interface LayoutCheckContext {
  image?: string | undefined;
  distance?: number | undefined;
  expectedScreenshotId?: string | undefined;
  actualScreenshotId?: string | undefined;
  distanceScreenshotId?: string | undefined;
  errorContextId?: string | undefined;
}
// ...
export interface StackTraceCause {
  className?: string | undefined;
  message?:
    | string
    | undefined;
  /** StackTraceCause cause = 4 [deprecated = true]; */
  stackTraceElements?: string[] | undefined;
}

export interface ScriptSource {
  fileName?: string | undefined;
  methodName?: string | undefined;
  lines?: ScriptSourceLine[] | undefined;
  mark?: number | undefined;
}

export interface ScriptSourceLine {
  line?:
    | string
    | undefined;
  /** bool mark = 3 [deprecated = true]; */
  lineNumber?: number | undefined;
}
// ...
export const ErrorContext: MessageFns<ErrorContext> = {
  encode(message: ErrorContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.scriptSource !== undefined) {
      ScriptSource.encode(message.scriptSource, writer.uint32(58).fork()).join();
    }
    if (message.ticketId !== undefined && message.ticketId !== "") {
      writer.uint32(74).string(message.ticketId);
    }
    if (message.description !== undefined && message.description !== "") {
      writer.uint32(82).string(message.description);
    }
    if (message.stackTrace !== undefined && message.stackTrace.length !== 0) {
      for (const v of message.stackTrace) {
        StackTraceCause.encode(v!, writer.uint32(90).fork()).join();
      }
    }
    if (message.optional !== undefined && message.optional !== false) {
      writer.uint32(96).bool(message.optional);
    }
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(106).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ErrorContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseErrorContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.scriptSource = ScriptSource.decode(reader, reader.uint32());
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.ticketId = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 11: {
          if (tag !== 90) {
            break;
          }

          const el = StackTraceCause.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.stackTrace!.push(el);
          }
          continue;
        }
        case 12: {
          if (tag !== 96) {
            break;
          }

          message.optional = reader.bool();
          continue;
        }
        case 13: {
          if (tag !== 106) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const LayoutCheckContext: MessageFns<LayoutCheckContext> = {
  encode(message: LayoutCheckContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.image !== undefined && message.image !== "") {
      writer.uint32(10).string(message.image);
    }
    if (message.distance !== undefined && message.distance !== 0) {
      writer.uint32(17).double(message.distance);
    }
    if (message.expectedScreenshotId !== undefined && message.expectedScreenshotId !== "") {
      writer.uint32(26).string(message.expectedScreenshotId);
    }
    if (message.actualScreenshotId !== undefined && message.actualScreenshotId !== "") {
      writer.uint32(34).string(message.actualScreenshotId);
    }
    if (message.distanceScreenshotId !== undefined && message.distanceScreenshotId !== "") {
      writer.uint32(42).string(message.distanceScreenshotId);
    }
    if (message.errorContextId !== undefined && message.errorContextId !== "") {
      writer.uint32(50).string(message.errorContextId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LayoutCheckContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLayoutCheckContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.image = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 17) {
            break;
          }

          message.distance = reader.double();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.expectedScreenshotId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.actualScreenshotId = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.distanceScreenshotId = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.errorContextId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const StackTraceCause: MessageFns<StackTraceCause> = {
  encode(message: StackTraceCause, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.className !== undefined && message.className !== "") {
      writer.uint32(10).string(message.className);
    }
    if (message.message !== undefined && message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    if (message.stackTraceElements !== undefined && message.stackTraceElements.length !== 0) {
      for (const v of message.stackTraceElements) {
        writer.uint32(26).string(v!);
      }
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): StackTraceCause {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStackTraceCause();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.className = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          const el = reader.string();
          if (el !== undefined) {
            message.stackTraceElements!.push(el);
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseScriptSource(): ScriptSource {
  return { fileName: "", methodName: "", lines: [], mark: 0 };
}

export const ScriptSource: MessageFns<ScriptSource> = {
  encode(message: ScriptSource, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.fileName !== undefined && message.fileName !== "") {
      writer.uint32(10).string(message.fileName);
    }
    if (message.methodName !== undefined && message.methodName !== "") {
      writer.uint32(18).string(message.methodName);
    }
    if (message.lines !== undefined && message.lines.length !== 0) {
      for (const v of message.lines) {
        ScriptSourceLine.encode(v!, writer.uint32(26).fork()).join();
      }
    }
    if (message.mark !== undefined && message.mark !== 0) {
      writer.uint32(32).int32(message.mark);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ScriptSource {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseScriptSource();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.fileName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.methodName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          const el = ScriptSourceLine.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.lines!.push(el);
          }
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.mark = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseScriptSourceLine(): ScriptSourceLine {
  return { line: "", lineNumber: 0 };
}

export const ScriptSourceLine: MessageFns<ScriptSourceLine> = {
  encode(message: ScriptSourceLine, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.line !== undefined && message.line !== "") {
      writer.uint32(10).string(message.line);
    }
    if (message.lineNumber !== undefined && message.lineNumber !== 0) {
      writer.uint32(16).int32(message.lineNumber);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ScriptSourceLine {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseScriptSourceLine();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.line = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.lineNumber = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}